+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- New function `uef_time:today/0`
+- New function `uef_time:tomorrow/0`
+- New function `uef_time:yesterday/0`
+
 ## [1.4.0] - 2019-05-11
 
 ### Added